+++ packages/form-render/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gopowerteam/form-render
 
+## 0.0.60
+
+### Patch Changes
+
+- Updated dependencies
+  - @gopowerteam/modal-render@0.0.39
+
 ## 0.0.59
 
 ### Patch Changes